requirements.txt
protobuf==3.20.3
openai
Pillow
python-dotenv
aiohttp
openai
google-genai
requests
google
google-api-core
google-generativeai
pysocks
speechkit
python-dotenv